+++ OLT.Core.sln
@@ -205,7 +205,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "OLT.Core.CommandBus", "OLT.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OLT.Core.CommandBus", "src\OLT.Core.CommandBus\OLT.Core.CommandBus.csproj", "{043F2D15-646E-4541-9995-E85C93A6896D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OLT.Core.CommandBus.Tests", "tests\OLT.Core.CommandBus.Tests\OLT.Core.CommandBus.Tests.csproj", "{391B2EDF-30C7-44C1-AAC8-582F45FDCD6C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OLT.Core.CommandBus.Tests", "tests\OLT.Core.CommandBus.Tests\OLT.Core.CommandBus.Tests.csproj", "{391B2EDF-30C7-44C1-AAC8-582F45FDCD6C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "OLT.AspNetCore.Versioning", "OLT.AspNetCore.Versioning", "{3E11448E-5ACC-46B8-AD23-3DD96C7F94ED}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OLT.AspNetCore.Versioning", "src\OLT.AspNetCore.Versioning\OLT.AspNetCore.Versioning.csproj", "{828461F3-7249-461E-A72D-DAFECFD555E3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -489,6 +493,10 @@ Global
 		{391B2EDF-30C7-44C1-AAC8-582F45FDCD6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{391B2EDF-30C7-44C1-AAC8-582F45FDCD6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{391B2EDF-30C7-44C1-AAC8-582F45FDCD6C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{828461F3-7249-461E-A72D-DAFECFD555E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{828461F3-7249-461E-A72D-DAFECFD555E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{828461F3-7249-461E-A72D-DAFECFD555E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{828461F3-7249-461E-A72D-DAFECFD555E3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -578,6 +586,8 @@ Global
 		{E5D44C0E-390D-4480-AF8A-E6968AB37300} = {1AA26E80-A124-4553-AE24-2721D851AB46}
 		{043F2D15-646E-4541-9995-E85C93A6896D} = {79FEE994-9D2D-43C1-A223-11078DB0F617}
 		{391B2EDF-30C7-44C1-AAC8-582F45FDCD6C} = {79FEE994-9D2D-43C1-A223-11078DB0F617}
+		{3E11448E-5ACC-46B8-AD23-3DD96C7F94ED} = {FC12AA4E-E1A3-4893-95F0-2230525468F2}
+		{828461F3-7249-461E-A72D-DAFECFD555E3} = {3E11448E-5ACC-46B8-AD23-3DD96C7F94ED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {94B70EDC-6E19-4295-BFA0-9474A97DF56F}
