+++ src/OLT.AspNetCore.Versioning/OltAspNetCoreVersioningDefaults.cs
@@ -1,0 +1,18 @@
+﻿namespace OLT.Constants
+{
+    public static class OltAspNetCoreVersioningDefaults
+    {       
+
+        public static class ApiVersion
+        {
+            public static class ParameterName
+            {
+                public const string Query = "api-version";
+                public const string MediaType = "v";
+                public const string Header = "x-api-version";
+            }
+        }
+
+
+    }
+}
