+++ src/OLT.AspNetCore.Versioning/Options/OltOptionsApiVersionParameter.cs
@@ -10,25 +10,25 @@ namespace OLT.Core
         /// Reads the API Version from the query string <see cref="MediaTypeApiVersionReader"/>
         /// </summary>
         /// <remarks>
-        /// Default <see cref="OltAspNetCoreDefaults.ApiVersion.ParameterName.Query"/>
+        /// Default <see cref="OltAspNetCoreVersioningDefaults.ApiVersion.ParameterName.Query"/>
         /// </remarks>
-        public virtual string Query { get; set; } = OltAspNetCoreDefaults.ApiVersion.ParameterName.Query;
+        public virtual string Query { get; set; } = OltAspNetCoreVersioningDefaults.ApiVersion.ParameterName.Query;
 
         /// <summary>
         /// Reads the API Version from media type <see cref="MediaTypeApiVersionReader"/>
         /// </summary>
         /// <remarks>
-        /// Default <see cref="OltAspNetCoreDefaults.ApiVersion.ParameterName.MediaType"/>
+        /// Default <see cref="OltAspNetCoreVersioningDefaults.ApiVersion.ParameterName.MediaType"/>
         /// </remarks>
-        public virtual string MediaType { get; set; } = OltAspNetCoreDefaults.ApiVersion.ParameterName.MediaType;
+        public virtual string MediaType { get; set; } = OltAspNetCoreVersioningDefaults.ApiVersion.ParameterName.MediaType;
 
         /// <summary>
         /// Reads the API Version from media type <see cref="MediaTypeApiVersionReader"/>
         /// </summary>
         /// <remarks>
-        /// Default <see cref="OltAspNetCoreDefaults.ApiVersion.ParameterName.Header"/>
+        /// Default <see cref="OltAspNetCoreVersioningDefaults.ApiVersion.ParameterName.Header"/>
         /// </remarks>
-        public virtual string Header { get; set; } = OltAspNetCoreDefaults.ApiVersion.ParameterName.Header;
+        public virtual string Header { get; set; } = OltAspNetCoreVersioningDefaults.ApiVersion.ParameterName.Header;
 
 
 
@@ -59,7 +59,7 @@ namespace OLT.Core
             }
             else
             {
-                readers.Add(new HeaderApiVersionReader(OltAspNetCoreDefaults.ApiVersion.ParameterName.Header));
+                readers.Add(new HeaderApiVersionReader(OltAspNetCoreVersioningDefaults.ApiVersion.ParameterName.Header));
             }
 
             readers.Add(new UrlSegmentApiVersionReader());
