+++ tests/OLT.AspNetCore.Versioning.Tests/UnitTest1.cs
@@ -1,0 +1,11 @@
+namespace OLT.AspNetCore.Versioning.Tests
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
